# spec/openapi.yml
...
openapi: 3.0.2
info:
  title: Marquez
  version: 0.51.1
  description: Marquez is an open source **metadata service** for the **collection**, **aggregation**,
    and **visualization** of a data ecosystem's metadata.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local API server
paths:
...
  /sources/{source}:
    parameters:
      - $ref: '#/components/parameters/source'
    put:
      operationId: putSource
      summary: Create a source
      description: Creates a new source object. A source is the physical location of a dataset such as
        a table in PostgreSQL, or topic in Kafka. A source enables the grouping of physical datasets
        to their physical source.
      deprecated: true
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedSource'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
    get:
      operationId: getSource
      summary: Retrieve a source
      description: Returns a source.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
  /sources:
    get:
      operationId: getSources
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all sources
      description: Returns a list of sources.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceList'
...
  /namespaces/{namespace}/jobs/{job}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
    put:
      operationId: putJob
      summary: Create a job
      description: Creates a new job object. All job objects are immutable and are uniquely identified by a generated ID.
        Marquez will create a version of a job each time the contents of the object is modified. For example, the `location`
        of a job may change over time resulting in new versions. The accumulated versions can be listed, used to rerun a
        specific job version or possibly help debug a failed job run.
      deprecated: true
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedJob'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
    get:
      operationId: getJob
      summary: Retrieve a job
      description: Retrieve a job.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
    delete:
      operationId: deleteJob
      summary: Soft deletes job.
      description: Soft deletes job. It will be un-deleted if new OpenLineage event containing this job comes.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
  /namespaces/{namespace}/jobs:
    parameters:
      - $ref: '#/components/parameters/namespace'
    get:
      operationId: getJobs
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all jobs
      description: Returns a list of jobs.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobList'
  /namespaces/{namespace}/jobs/{job}/versions/{version}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
      - $ref: '#/components/parameters/version'
    get:
      operationId: getJobVersion
      summary: Retrieve a version for a job
      description: Returns a version for a job.
      tags:
        - Jobs
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobVersion'
  /namespaces/{namespace}/jobs/{job}/versions:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
    get:
      operationId: getJobVersions
      summary: List all versions for a job
      description: Returns a list of versions for a job.
      tags:
        - Jobs
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobVersionList'
  /namespaces/{namespace}/jobs/{job}/runs:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/job'
    post:
      operationId: createRun
      summary: Create a run
      description: Creates a new run object for a job.
      deprecated: true
      tags:
        - Jobs
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedRun'
      responses:
        '201':
          description: CREATED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
    get:
      operationId: getRuns
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all runs
      description: Returns a list of runs for a job.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
  /jobs/runs/{id}:
    parameters:
      - $ref: '#/components/parameters/runId'
    get:
      operationId: getRun
      summary: Retrieve a run
      description: Retrieve a run.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
  /jobs/runs/{id}/facets:
    parameters:
      - $ref: '#/components/parameters/runId'
      - $ref: '#/components/parameters/type'
    get:
      operationId: getFacets
      summary: Retrieve run or job facets for a run.
      description: Retrieve run or job facets for a run.
      tags:
        - Jobs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetsResult'
...
components:
  parameters:
...
    namespace:
      name: namespace
      in: path
      description: The name of the namespace.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-namespace
    source:
      name: source
      in: path
      description: The name of the source.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-source
...
    job:
      name: job
      in: path
      description: The name of the job.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-job
    runId:
      name: id
      in: path
      description: The ID of the run.
      required: true
      schema:
        type: string
        format: uuid
        example: ea9badc5-7cb2-49af-9a9f-155771d3a797
    version:
      name: version
      in: path
      description: The ID of the job or dataset version.
      required: true
      schema:
        type: string
        format: uuid
        example: ea9badc5-7cb2-49af-9a9f-155771d3a797
    limit:
      name: limit
      in: query
      description: The number of results to return from offset.
      required: false
      schema:
        type: integer
        example: 25
        default: 100
    offset:
      name: offset
      in: query
      description: The initial position from which to return results.
      required: false
      schema:
        type: integer
        default: 0
...
    type:
      name: type
      in: query
      description: Indicates if should return job or run facets.
      required: true
      schema:
        type: string
        enum:
          - run
          - job
  schemas:
...
    CreatedSource:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the source.
          type: string
      required:
        - type
        - connectionUrl
      example:
        type: POSTGRESQL
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    Source:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        name:
          description: The name of the source.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was updated.
          type: string
          format: date-time
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the dataset.
          type: string
      example:
        type: POSTGRESQL
        name: my-source
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    SourceList:
      type: object
      properties:
        sources:
          items:
            $ref: '#/components/schemas/Source'
...
    DatasetId:
      type: object
      description: The ID of the dataset.
      properties:
        namespace:
          type: string
          description: The namespace of the dataset.
        name:
          type: string
          description: The name of the dataset.
...
    JobType:
      type: string
      enum: [BATCH, STREAM, SERVICE]
      description: The type of the job.
    CreatedJob:
      type: object
      properties:
        id:
          description: The ID of the job.
          type: object
          properties:
            namespace:
              type: string
              description: The namespace of the job.
            name:
              type: string
              description: The name of the job.
        type:
          $ref: '#/components/schemas/JobType'
        inputs:
          description: The set of **input** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        outputs:
          description: The set of **output** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        location:
          description: The URL of the job source code or artifact.
          type: string
          format: URL
        context:
          description: A key/value pair that must be of type `string`. A context can be used for getting additional details about the job.
          type: object
          deprecated: true
        description:
          description: The description of the job.
          type: string
        runId:
          description: An _optional_ run ID used to associate a job version to an existing job run.
          type: string
      required:
        - type
        - inputs
        - outputs
      example:
        type: BATCH
        inputs: [{ namespace: my-namespace, name: my-dataset}]
        outputs: []
        location: https://github.com/my-jobs/blob/124f6089ad4c5fcbb1d7b33cbb5d3a9521c5d32c
        context: {sql: 'SELECT * FROM mytable;'}
        description: My first job!
    JobId:
      description: The ID of the job.
      type: object
      properties:
        namespace:
          type: string
          description: The namespace of the job.
        name:
          type: string
          description: The name of the job.
    Job:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/JobId'
        type:
          description: The type of the job.
          type: string
        name:
          description: The name of the job.
          type: string
        simpleName:
          description: The shortened name of the job.
          type: string
        parentJobName:
          description: The name of the parent job.
          type: string
        parentJobUuid:
          description: The UUID of the parent job.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the job was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the job was updated.
          type: string
          format: date-time
        namespace:
          description: The namespace of the job.
          type: string
        inputs:
          description: The set of **input** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        outputs:
          description: The set of **output** datasets.
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        location:
          description: The URL of the job source code or artifact.
          type: string
          format: URL
        context:
          description: A key/value pair that must be of type `string`. A context can be used for getting additional details about the job.
          type: object
          deprecated: true
        description:
          description: The description of the job.
          type: string
        latestRun:
          $ref: '#/components/schemas/Run'
        latestRuns:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        facets:
          $ref: '#/components/schemas/JobFacets'
        currentVersion:
          description: The current version of the job.
          type: string
          format: uuid
      example:
        id:
          namespace: my-namespace
          name: my-job
        type: BATCH
        name: my-job
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        namespace: my-namespace
        inputs: [{namespace: my-namespace, name: my-dataset}]
        outputs: []
        location: https://github.com/my-jobs/blob/124f6089ad4c5fcbb1d7b33cbb5d3a9521c5d32c
        context: {'SQL': "SELECT * FROM mytable;"}
        description: My first job!
        latestRun: null
        latestRuns: []
        facets: {}
        currentVersion: "b1d626a2-6d3a-475e-9ecf-943176d4a8c6"
    JobVersionList:
      type: object
      properties:
        versions:
          items:
            $ref: '#/components/schemas/JobVersion'
    JobVersionId:
      type: object
      properties:
        namespace:
          type: string
          description: The namespace of the job.
        name:
          type: string
          description: The name of the job.
        version:
          type: string
          format: uuid
          description: The version of the job.
    JobVersion:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/JobVersionId'
        name:
          description: The name of the job.
          type: string
        version:
          description: The version of the job
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the job version was created.
          type: string
          format: date-time
        namespace:
          description: The namespace of the job.
          type: string
        inputs:
          description: The set of **input** datasets
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        outputs:
          description: The set of **output** datasets
          type: array
          items:
            $ref: '#/components/schemas/DatasetId'
          uniqueItems: true
        location:
          description: The URL of the job source code or artifact for the given job version.
          type: string
          format: URL
        context:
          description: Deprecated- Use facet. A key/value pair that must be of type `string`. A context can be used for getting additional details about the job.
          type: object
          deprecated: true
        latestRun:
          $ref: '#/components/schemas/Run'
        facets:
          $ref: '#/components/schemas/JobFacets'
      example:
        id:
          namespace: my-namespace
          name: my-job
          version: 56472c57-a2ef-4218-b7b7-d2af02a343fd
        type: BATCH
        name: my-job
        version: 56472c57-a2ef-4218-b7b7-d2af02a343fd
        createdAt: 2019-05-09T19:49:24.201361Z
        namespace: my-namespace
        inputs: [{namespace: my-namespace, name: my-dataset}]
        outputs: []
        location: https://github.com/my-jobs/blob/124f6089ad4c5fcbb1d7b33cbb5d3a9521c5d32c
        context: {'SQL': "SELECT * FROM mytable;"}
        description: My first job!
        facets: {}
    JobList:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        totalCount:
          type: number
          description: The total number of jobs for the given namespace
    CreatedRun:
      type: object
      properties:
        id:
          description: An _optional_ user-provided unique ID of the run. A run ID **must** be an [UUID](https://tools.ietf.org/html/rfc4122).
            If an ID for the run is not provided, a random UUID will be generated for the given run.
          type: string
          format: uuid
        nominalStartTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal start time of the run.
          type: string
          format: date-time
        nominalEndTime:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the nominal end time of the run.
          type: string
          format: date-time
        args:
          description: The arguments of the run.
          type: object
      example:
        args: {'email': 'me@example.com', 'emailOnFailure': false, 'emailOnRetry': true, 'retries': 1}
...
    Run:
      type: object
      anyOf:
        - $ref: '#/components/schemas/IncompleteRun'
        - type: object
          properties:
            jobVersion:
              $ref: '#/components/schemas/JobVersionId'
            inputDatasetVersions:
              type: array
              items:
                $ref: '#/components/schemas/InputDatasetVersion'
            outputDatasetVersions:
              type: array
              items:
                $ref: '#/components/schemas/OutputDatasetVersion'
            context:
              description: A key/value pair that must be of type `string`. A context can be used for getting additional details about the job.
              type: object
              deprecated: true
            facets:
              $ref: '#/components/schemas/RunFacets'
      example:
        id: 870492da-ecfb-4be0-91b9-9a89ddd3db90
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        nominalStartTime: null
        nominalEndTime: null
        state: COMPLETED
        startedAt: 2019-05-09T15:17:32.690346
        endedAt: 2019-05-09T20:05:46.815920Z
        durationMs: 4250894125
        args: {'email': 'me@example.com', 'emailOnFailure': 'false', 'emailOnRetry': 'true', 'retries': '1'}
        context: {'SQL': "SELECT * FROM mytable;"}
        facets: {}
    RunList:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/Run'
...
    JobFacets:
      description: The facets of the job. For a full list of standard job facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
    RunFacets:
      description: The facets of the run. For a full list of standard run facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
    CustomFacet:
      description: A custom facet enables the extension of _dataset_, _job_, and _run_ metadata. A custom facet **must** also have a schema,
        where a version of the schema is identifiable via a URL. A field within the schema **must** not start with an underscore (`_`).
      allOf:
        - $ref: '#/components/schemas/BaseFacet'
        - type: object
          additionalProperties: true
...
    FacetsResult:
      type: object
      properties:
        runId:
          description: The ID associated with the run modifying the table.
          type: string
        facets:
          $ref: '#/components/schemas/CustomFacet'
    InputDatasetVersion:
      type: object
      properties:
        datasetVersionId:
          $ref: '#/components/schemas/DatasetVersionId'
        facets:
          type: object
          additionalProperties:
            type: string
          description: Dataset facets in run context, like `inputFacets`.
      required:
        - datasetVersionId
        - facets
    OutputDatasetVersion:
      type: object
      properties:
        datasetVersionId:
          $ref: '#/components/schemas/DatasetVersionId'
        facets:
          type: object
          additionalProperties:
            type: string
          description: Dataset facets in run context, like `outputFacets`.
      required:
        - datasetVersionId
        - facets